+++ .env
@@ -1,8 +1,8 @@
-VITE_APP_VERSION = v1.4.0
-GENERATE_SOURCEMAP = false
+VITE_APP_VERSION=v1.4.0
+GENERATE_SOURCEMAP= alse
 
 ## Public URL
-VITE_APP_BASE_NAME = /mts
+VITE_APP_BASE_NAME=/mts
 
-VITE_API_BASE_URL = https://looking-erp-cheats-twelve.trycloudflare.com/api
+VITE_API_BASE_URL=https://korean-governor-amended-sh.trycloudflare.com/api
