+++ src/presentation/components/layout/Dashboard/Header/index.jsx
@@ -10,7 +10,7 @@ import AppBarStyled from './AppBarStyled';
 import HeaderContent from './HeaderContent';
 import IconButton from '../../../@extended/IconButton';
 
-import { handlerDrawerOpen, useGetMenuMaster } from '../../../../../infrastructure/api/http/menu';
+import { handlerDrawerOpen, useGetMenuMaster } from '../../../../../infrastructure/utils/menu';
 import { DRAWER_WIDTH, MINI_DRAWER_WIDTH } from '../../../../../app/constants';
 
 // assets
