+++ src/presentation/components/logo/LogoIcon.jsx
@@ -1,5 +1,5 @@
 import Logo from '../../assets/images/logo/logo.png';
 
-export default function LogoIcon() {
-  return <img src={Logo} width={90} height={90} />;
+export default function LogoIcon({ width = 90, height = 90 }) {
+  return <img src={Logo} width={width} height={height} />;
 }
